+++ sharding-jdbc-core/src/main/java/io/shardingjdbc/core/parsing/lexer/dialect/oracle/OracleKeyword.java
@@ -27,7 +27,6 @@ import io.shardingjdbc.core.parsing.lexer.token.Keyword;
 public enum OracleKeyword implements Keyword {
     
     LOCKED,
-    COMMIT,
     CREATION,
     UPDATED,
     UPSERT,
@@ -36,7 +35,6 @@ public enum OracleKeyword implements Keyword {
     MERGE,
     PURGE,
     ROWS,
-    IF,
     GOTO,
     ONLY,
     AUTOMATIC,
@@ -76,8 +74,6 @@ public enum OracleKeyword implements Keyword {
     MAXTRANS,
     MINEXTENTS,
     MAXEXTENTS,
-    BEGIN,
-    SAVEPOINT,
     MATCHED,
     LOB,
     DIMENSION,