+++ sharding-jdbc-core/src/main/java/io/shardingjdbc/core/parsing/lexer/dialect/sqlserver/SQLServerKeyword.java
@@ -35,7 +35,6 @@ public enum SQLServerKeyword implements Keyword {
     ONLY, 
     OUTPUT, 
     AUTO, 
-    BEGIN, 
     IDENTITY, 
     BROWSE, 
     TYPE, 
@@ -48,5 +47,6 @@ public enum SQLServerKeyword implements Keyword {
     APPLY,
     REDUCE,
     REPLICATE,
+    EXTRACT,
     REDISTRIBUTE
 }