+++ sharding-jdbc-core/src/main/java/io/shardingjdbc/core/parsing/parser/clause/OrderByClauseParser.java
@@ -2,9 +2,12 @@ package io.shardingjdbc.core.parsing.parser.clause;
 
 import io.shardingjdbc.core.constant.OrderType;
 import io.shardingjdbc.core.parsing.lexer.LexerEngine;
+import io.shardingjdbc.core.parsing.lexer.dialect.oracle.OracleKeyword;
 import io.shardingjdbc.core.parsing.lexer.token.DefaultKeyword;
 import io.shardingjdbc.core.parsing.lexer.token.Symbol;
+import io.shardingjdbc.core.parsing.parser.clause.expression.BasicExpressionParser;
 import io.shardingjdbc.core.parsing.parser.context.OrderItem;
+import io.shardingjdbc.core.parsing.parser.dialect.ExpressionParserFactory;
 import io.shardingjdbc.core.parsing.parser.exception.SQLParsingException;
 import io.shardingjdbc.core.parsing.parser.expression.SQLExpression;
 import io.shardingjdbc.core.parsing.parser.expression.SQLIdentifierExpression;
@@ -28,11 +31,11 @@ public class OrderByClauseParser implements SQLClauseParser {
     @Getter
     private final LexerEngine lexerEngine;
     
-    private final ExpressionClauseParser expressionClauseParser;
+    private final BasicExpressionParser basicExpressionParser;
     
     public OrderByClauseParser(final LexerEngine lexerEngine) {
         this.lexerEngine = lexerEngine;
-        expressionClauseParser = new ExpressionClauseParser(lexerEngine);
+        basicExpressionParser = ExpressionParserFactory.createBasicExpressionParser(lexerEngine);
     }
     
     /**
@@ -45,7 +48,7 @@ public class OrderByClauseParser implements SQLClauseParser {
             return;
         }
         List<OrderItem> result = new LinkedList<>();
-        lexerEngine.skipIfEqual(DefaultKeyword.SIBLINGS);
+        lexerEngine.skipIfEqual(OracleKeyword.SIBLINGS);
         lexerEngine.accept(DefaultKeyword.BY);
         do {
             result.add(parseSelectOrderByItem(selectStatement));
@@ -55,7 +58,7 @@ public class OrderByClauseParser implements SQLClauseParser {
     }
     
     private OrderItem parseSelectOrderByItem(final SelectStatement selectStatement) {
-        SQLExpression sqlExpression = expressionClauseParser.parse(selectStatement);
+        SQLExpression sqlExpression = basicExpressionParser.parse(selectStatement);
         OrderType orderByType = OrderType.ASC;
         if (lexerEngine.skipIfEqual(DefaultKeyword.ASC)) {
             orderByType = OrderType.ASC;