+++ sharding-jdbc-core/src/main/java/io/shardingjdbc/core/rewrite/SQLBuilder.java
@@ -17,6 +17,7 @@
 
 package io.shardingjdbc.core.rewrite;
 
+import com.google.common.base.Strings;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 
@@ -56,6 +57,18 @@ public final class SQLBuilder {
     }
     
     /**
+     * Append index token.
+     *
+     * @param indexName index name
+     * @param tableName table name
+     */
+    public void appendIndex(final String indexName, final String tableName) {
+        segments.add(new IndexToken(indexName, tableName));
+        currentSegment = new StringBuilder();
+        segments.add(currentSegment);
+    }
+    
+    /**
      * Append table token.
      *
      * @param tableName table name
@@ -77,6 +90,14 @@ public final class SQLBuilder {
         for (Object each : segments) {
             if (each instanceof TableToken && tableTokens.containsKey(((TableToken) each).tableName)) {
                 result.append(tableTokens.get(((TableToken) each).tableName));
+            } else if (each instanceof IndexToken) {
+                IndexToken indexToken = (IndexToken) each;
+                result.append(indexToken.indexName);
+                String tableName = tableTokens.get(indexToken.tableName);
+                if (!Strings.isNullOrEmpty(tableName)) {
+                    result.append("_");
+                    result.append(tableName);
+                }
             } else {
                 result.append(each);
             }
@@ -92,6 +113,19 @@ public final class SQLBuilder {
         @Override
         public String toString() {
             return tableName;
+        }
+    }
+    
+    @RequiredArgsConstructor
+    private class IndexToken {
+        
+        private final String indexName;
+        
+        private final String tableName;
+        
+        @Override
+        public String toString() {
+            return indexName;
         }
     }
 }