+++ sharding-jdbc-core/src/main/java/io/shardingjdbc/core/rewrite/SQLRewriteEngine.java
@@ -17,15 +17,15 @@
 
 package io.shardingjdbc.core.rewrite;
 
-
+import com.google.common.base.Optional;
+import com.google.common.base.Strings;
 import io.shardingjdbc.core.constant.DatabaseType;
-import io.shardingjdbc.core.rule.BindingTableRule;
-import io.shardingjdbc.core.rule.ShardingRule;
 import io.shardingjdbc.core.parsing.lexer.token.DefaultKeyword;
 import io.shardingjdbc.core.parsing.parser.context.OrderItem;
 import io.shardingjdbc.core.parsing.parser.context.limit.Limit;
 import io.shardingjdbc.core.parsing.parser.sql.SQLStatement;
 import io.shardingjdbc.core.parsing.parser.sql.dql.select.SelectStatement;
+import io.shardingjdbc.core.parsing.parser.token.IndexToken;
 import io.shardingjdbc.core.parsing.parser.token.ItemsToken;
 import io.shardingjdbc.core.parsing.parser.token.OffsetToken;
 import io.shardingjdbc.core.parsing.parser.token.OrderByToken;
@@ -34,7 +34,8 @@ import io.shardingjdbc.core.parsing.parser.token.SQLToken;
 import io.shardingjdbc.core.parsing.parser.token.TableToken;
 import io.shardingjdbc.core.routing.type.TableUnit;
 import io.shardingjdbc.core.routing.type.complex.CartesianTableReference;
-import com.google.common.base.Optional;
+import io.shardingjdbc.core.rule.BindingTableRule;
+import io.shardingjdbc.core.rule.ShardingRule;
 import io.shardingjdbc.core.util.SQLUtil;
 
 import java.util.Collections;
@@ -99,6 +100,8 @@ public final class SQLRewriteEngine {
             }
             if (each instanceof TableToken) {
                 appendTableToken(result, (TableToken) each, count, sqlTokens);
+            } else if (each instanceof IndexToken) {
+                appendIndexToken(result, (IndexToken) each, count, sqlTokens);
             } else if (each instanceof ItemsToken) {
                 appendItemsToken(result, (ItemsToken) each, count, sqlTokens);
             } else if (each instanceof RowCountToken) {
@@ -124,10 +127,19 @@ public final class SQLRewriteEngine {
     }
     
     private void appendTableToken(final SQLBuilder sqlBuilder, final TableToken tableToken, final int count, final List<SQLToken> sqlTokens) {
-        String tableName = 
-                SQLUtil.getOriginalValue(sqlStatement.getTables().getTableNames().contains(tableToken.getTableName()) ? tableToken.getTableName() : tableToken.getOriginalLiterals(), databaseType);
-        sqlBuilder.appendTable(tableName);
+        sqlBuilder.appendTable(tableToken.getTableName().toLowerCase());
         int beginPosition = tableToken.getBeginPosition() + tableToken.getOriginalLiterals().length();
+        appendRest(sqlBuilder, count, sqlTokens, beginPosition);
+    }
+    
+    private void appendIndexToken(final SQLBuilder sqlBuilder, final IndexToken indexToken, final int count, final List<SQLToken> sqlTokens) {
+        String indexName = indexToken.getIndexName().toLowerCase();
+        String logicTableName = indexToken.getTableName().toLowerCase();
+        if (Strings.isNullOrEmpty(logicTableName)) {
+            logicTableName = shardingRule.getLogicTableName(indexName);
+        }
+        sqlBuilder.appendIndex(indexName, logicTableName);
+        int beginPosition = indexToken.getBeginPosition() + indexToken.getOriginalLiterals().length();
         appendRest(sqlBuilder, count, sqlTokens, beginPosition);
     }
     
@@ -208,9 +220,10 @@ public final class SQLRewriteEngine {
     }
     
     private Map<String, String> getTableTokens(final TableUnit tableUnit) {
+        String logicTableName = tableUnit.getLogicTableName().toLowerCase();
         Map<String, String> tableTokens = new HashMap<>();
-        tableTokens.put(tableUnit.getLogicTableName(), tableUnit.getActualTableName());
-        Optional<BindingTableRule> bindingTableRule = shardingRule.findBindingTableRule(tableUnit.getLogicTableName());
+        tableTokens.put(logicTableName, tableUnit.getActualTableName());
+        Optional<BindingTableRule> bindingTableRule = shardingRule.findBindingTableRule(logicTableName);
         if (bindingTableRule.isPresent()) {
             tableTokens.putAll(getBindingTableTokens(tableUnit, bindingTableRule.get()));
         }
@@ -220,8 +233,9 @@ public final class SQLRewriteEngine {
     private Map<String, String> getTableTokens(final CartesianTableReference cartesianTableReference) {
         Map<String, String> tableTokens = new HashMap<>();
         for (TableUnit each : cartesianTableReference.getTableUnits()) {
-            tableTokens.put(each.getLogicTableName(), each.getActualTableName());
-            Optional<BindingTableRule> bindingTableRule = shardingRule.findBindingTableRule(each.getLogicTableName());
+            String logicTableName = each.getLogicTableName().toLowerCase();
+            tableTokens.put(logicTableName, each.getActualTableName());
+            Optional<BindingTableRule> bindingTableRule = shardingRule.findBindingTableRule(logicTableName);
             if (bindingTableRule.isPresent()) {
                 tableTokens.putAll(getBindingTableTokens(each, bindingTableRule.get()));
             }
@@ -232,8 +246,9 @@ public final class SQLRewriteEngine {
     private Map<String, String> getBindingTableTokens(final TableUnit tableUnit, final BindingTableRule bindingTableRule) {
         Map<String, String> result = new HashMap<>();
         for (String eachTable : sqlStatement.getTables().getTableNames()) {
-            if (!eachTable.equalsIgnoreCase(tableUnit.getLogicTableName()) && bindingTableRule.hasLogicTable(eachTable)) {
-                result.put(eachTable, bindingTableRule.getBindingActualTable(tableUnit.getDataSourceName(), eachTable, tableUnit.getActualTableName()));
+            String tableName = eachTable.toLowerCase();
+            if (!tableName.equals(tableUnit.getLogicTableName().toLowerCase()) && bindingTableRule.hasLogicTable(tableName)) {
+                result.put(tableName, bindingTableRule.getBindingActualTable(tableUnit.getDataSourceName(), tableName, tableUnit.getActualTableName()));
             }
         }
         return result;