+++ sharding-jdbc-core/src/test/java/io/shardingjdbc/core/integrate/type/sharding/AbstractShardingTableOnlyTest.java
@@ -29,13 +29,9 @@ import io.shardingjdbc.core.integrate.fixture.PreciseModuloTableShardingAlgorith
 import io.shardingjdbc.core.integrate.fixture.RangeModuloTableShardingAlgorithm;
 import io.shardingjdbc.core.integrate.jaxb.SQLShardingRule;
 import io.shardingjdbc.core.jdbc.core.datasource.ShardingDataSource;
-import org.junit.After;
-import org.junit.Before;
 
 import javax.sql.DataSource;
-import java.sql.Connection;
 import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
@@ -71,6 +67,7 @@ public abstract class AbstractShardingTableOnlyTest extends AbstractSQLAssertTes
             final ShardingRuleConfiguration shardingRuleConfig = new ShardingRuleConfiguration();
             TableRuleConfiguration orderTableRuleConfig = new TableRuleConfiguration();
             orderTableRuleConfig.setLogicTable("t_order");
+            orderTableRuleConfig.setLogicIndex("t_order_index");
             List<String> orderActualDataNodes = new LinkedList<>();
             for (String dataSourceName : entry.getValue().keySet()) {
                 orderActualDataNodes.add(dataSourceName + ".t_order_${0..9}");
@@ -78,14 +75,21 @@ public abstract class AbstractShardingTableOnlyTest extends AbstractSQLAssertTes
             orderTableRuleConfig.setActualDataNodes(Joiner.on(",").join(orderActualDataNodes));
             shardingRuleConfig.getTableRuleConfigs().add(orderTableRuleConfig);
             TableRuleConfiguration orderItemTableRuleConfig = new TableRuleConfiguration();
-            orderItemTableRuleConfig.setLogicTable("t_order_item");
             List<String> orderItemActualDataNodes = new LinkedList<>();
             for (String dataSourceName : entry.getValue().keySet()) {
                 orderItemActualDataNodes.add(dataSourceName + ".t_order_item_${0..9}");
             }
             orderItemTableRuleConfig.setActualDataNodes(Joiner.on(",").join(orderItemActualDataNodes));
+            orderItemTableRuleConfig.setLogicTable("t_order_item");
             orderItemTableRuleConfig.setKeyGeneratorClass("item_id");
             shardingRuleConfig.getTableRuleConfigs().add(orderItemTableRuleConfig);
+            TableRuleConfiguration logTableRuleConfig = new TableRuleConfiguration();
+            logTableRuleConfig.setLogicIndex("t_log_index");
+            logTableRuleConfig.setLogicTable("t_log");
+            TableRuleConfiguration tempLogTableRuleConfig = new TableRuleConfiguration();
+            tempLogTableRuleConfig.setLogicTable("t_temp_log");
+            shardingRuleConfig.getTableRuleConfigs().add(logTableRuleConfig);
+            shardingRuleConfig.getTableRuleConfigs().add(tempLogTableRuleConfig);
             shardingRuleConfig.getBindingTableGroups().add("t_order, t_order_item");
             shardingRuleConfig.setDefaultDatabaseShardingStrategyConfig(new NoneShardingStrategyConfiguration());
             shardingRuleConfig.setDefaultTableShardingStrategyConfig(
@@ -93,43 +97,5 @@ public abstract class AbstractShardingTableOnlyTest extends AbstractSQLAssertTes
             getShardingDataSources().put(entry.getKey(), new ShardingDataSource(shardingRuleConfig.build(entry.getValue())));
         }
         return getShardingDataSources();
-    }
-    
-    @Before
-    public void initDDLTables() throws SQLException {
-        if (getSql().startsWith("ALTER") || getSql().startsWith("TRUNCATE") || getSql().startsWith("DROP")) {
-            if (getSql().contains("TEMP")) {
-                executeSql("CREATE TEMPORARY TABLE t_temp_log(id int, status varchar(10))");
-            } else {
-                executeSql("CREATE TABLE t_log(id int, status varchar(10))");
-            }
-        }
-    }
-    
-    @After
-    public void cleanupDdlTables() throws SQLException {
-        if (getSql().startsWith("CREATE") || getSql().startsWith("ALTER") || getSql().startsWith("TRUNCATE")) {
-            if (getSql().contains("TEMP")) {
-                executeSql("DROP TABLE t_temp_log");
-            } else {
-                executeSql("DROP TABLE t_log");
-            }
-        }
-    }
-    
-    private void executeSql(final String sql) throws SQLException {
-        for (Map.Entry<DatabaseType, ShardingDataSource> each : getDataSources().entrySet()) {
-            if (getCurrentDatabaseType() == each.getKey()) {
-                try (Connection conn = each.getValue().getConnection();
-                     Statement statement = conn.createStatement()) {
-                    statement.execute(sql);
-                    //CHECKSTYLE:OFF
-                } catch (final Exception ex) {
-                    //CHECKSTYLE:ON
-                    ex.printStackTrace();
-                    throw new RuntimeException(ex);
-                }
-            }
-        }
     }
 }