+++ sharding-jdbc-core/src/test/java/io/shardingjdbc/core/parsing/AllParsingTests.java
@@ -18,19 +18,19 @@
 package io.shardingjdbc.core.parsing;
 
 import io.shardingjdbc.core.parsing.lexer.AllLexerTests;
-import io.shardingjdbc.core.parsing.lexer.analyzer.TokenizerTest;
+import io.shardingjdbc.core.parsing.parser.context.OrderItemTest;
 import io.shardingjdbc.core.parsing.parser.sql.AllStatementParserTests;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
-        TokenizerTest.class,
         AllLexerTests.class,
         AllStatementParserTests.class,
         SQLParsingEngineTest.class,
         UnsupportedSQLParsingEngineTest.class,
-        SQLJudgeEngineTest.class
+        SQLJudgeEngineTest.class,
+        OrderItemTest.class
     })
 public class AllParsingTests {
 }